Added avx512 packed vector cross product

+++ book-01/Assembly/asm-headers/avx_packed.h
@@ -28,7 +28,7 @@ namespace Assembly {
 			__declspec(dllexport) void Y_Abs_Double(const YmmVal& input, YmmVal results[]);
 			__declspec(dllexport) bool Y_Array_Cols_Mean_Double(const double* input, uint32_t numRows, uint32_t numCols, double* colMeans);
 			__declspec(dllexport) void Y_Blend_Float(YmmVal * src1, YmmVal * src2, YmmVal * des, BlendSource blend[]);
-			__declspec(dllexport) bool Y_Calc_Correlation_Coefficient(const double* x, const double* y, size_t arrayLength, double sums[5], double epsilon, double* rho);
+			__declspec(dllexport) bool Y_Calc_Correlation_Coefficient(const double* x, const double* y, uint32_t arrayLength, double sums[5], double epsilon, double* rho);
 			__declspec(dllexport) void Y_Calc_Sphere_Area_Volume(const float* radius, uint32_t arrayLength, float* areas, float* volumes, float qnanValue);
 			__declspec(dllexport) void Y_Matrix_Multiplication_Double(const double* a, const double* b, double* result);
 			__declspec(dllexport) void Y_Matrix_Transpose_Double(double* input, const double* result);
